% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIC.R
\name{DIC}
\alias{DIC}
\title{Deviance Information Criteria (DIC)}
\usage{
DIC(m1, ...)
}
\arguments{
\item{m1}{MCMCglmm model}

\item{...}{Other MCMCglmm models}
}
\description{
Extracts DIC from a individual or set of MCMCglmm models. If a set of models is provided, provides a table of DIC differences between candidate models.
}
\examples{
DIC(mod1)
DIC(mod1,mod2)
}
