% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFixedTerms.R
\name{plotFixedTerms}
\alias{plotFixedTerms}
\title{95\% credible intervals for posterior distibution.}
\usage{
plotFixedTerms(mod1)
}
\arguments{
\item{mod1}{MCMCglmm model}
}
\description{
\code{plotFixedTerms} plots 95\% credible intervals for posterior distibution
using \code{ggplot2}
}
\examples{
plotFixedTerms(mod1)
}
