% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varComp.R
\name{varComp}
\alias{varComp}
\title{Variance components for the random effects of an MCMCglmm model.}
\usage{
varComp(mod, units = T)
}
\arguments{
\item{mod}{MCMCglmm model}

\item{units}{Use unit ("residual") variance?}
}
\description{
Calculates proportion of variance component taken up by each random effect. Returns posterior mode from kernel density estimation, as well as upper and lower CIs for each random effect.
}
\details{
NOTE: Use units=T with caution if running multinomial models (unit variance may be unestimable, and should be fixed at a constant).
}
\examples{
load(data)
mod1<-MCMCglmm(Count~x,random=~y+z,data=data,family='poisson')
varComp(mod1)
}
