% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zeroCheck.R
\name{zeroCheck}
\alias{zeroCheck}
\title{Expected zeros from a Poisson model}
\usage{
zeroCheck(mod, dat)
}
\arguments{
\item{dat}{Data frame used to fit \code{mod1}}

\item{mod1}{MCMCglmm model}
}
\description{
Posterior predictive checks for zero-inflation (MCMCglmm course notes pg. 104). Runs simulation NITT times, using individual realizations from the model to simulate a poisson distribution. The number of zeros from each simulation are plotted along with the actual number of zeros from the dataset, in order to decide if the model requires a zero-inflation term ('zipoisson')
}
\examples{
load(data)
mod1<-MCMCglmm(Count~x,data=data,family='poisson')
zeroCheck(mod1,data)
}
